Feat: [EVER-211] NotFound에서 Empty 컴포넌트 변경
