+++ src/pages/common/Empty.tsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import { Button } from '@/components/Button';
+import { useNavigate } from 'react-router-dom';
+
+interface EmptyProps {
+  imageSrc: string;
+  altText?: string;
+  message: string;
+  buttonText?: string;
+  buttonLink?: string;
+  className?: string;
+}
+
+const Empty: React.FC<EmptyProps> = ({
+  imageSrc,
+  altText = '',
+  message,
+  buttonText,
+  buttonLink,
+  className = '',
+}) => {
+  const navigate = useNavigate();
+
+  return (
+    <div className={`flex flex-col items-center justify-center py-12 ${className}`}>
+      <img src={imageSrc} alt={altText} className="w-40 h-auto mb-4" />
+      <p className="text-gray-500 text-sm text-center mb-4">{message}</p>
+
+      {buttonText && buttonLink && (
+        <Button
+          variant="outline"
+          size="default"
+          className="bg-white/50 border-black/50 text-gray-800 hover:bg-white/70"
+          onClick={() => navigate(buttonLink)}
+        >
+          {buttonText}
+        </Button>
+      )}
+    </div>
+  );
+};
+
+export default Empty;
